refactoring

+++ app/models/automatic_ban.rb
@@ -21,13 +21,8 @@ class AutomaticBan
   end
 
   def create
-    @unpaid_invoices_ids = Invoice.where(user_id: user, status: Invoice.statuses[:cancelled])
-                                  .pluck(:id)
-    @existing_bans_ids = Ban.where(invoice_id: @unpaid_invoices_ids).pluck(:id)
-
-    if @unpaid_invoices_ids.count == @existing_bans_ids.count
-      raise Errors::NoCancelledInvoices.new(user.id, domain_name)
-    end
+    generate_ids_for_invoices_and_bans
+    raise_no_cancelled_invoices
 
     create_ban
     assign_ban_length
@@ -57,9 +52,9 @@ class AutomaticBan
   def remove_offer_from_active_auction
     auction = Auction.find_by(domain_name: @domain_name)
 
-    return unless auction.present?
+    return if auction.blank?
     return unless auction.in_progress?
-    
+
     offer = @user.offers.find_by(auction_id: auction.id)
     offer.destroy if offer.present?
   end
@@ -67,11 +62,23 @@ class AutomaticBan
   def remove_offers_from_active_auctions
     ban_number_of_strikes = Setting.find_by(code: 'ban_number_of_strikes')
     return unless ban_number_of_strikes.value.to_i <= @user.bans.count
-    
+
     @user.offers.each do |offer|
       auction_id = offer.auction_id
       auction = Auction.find(auction_id)
-      offer.destroy
+      offer.destroy if auction.in_progress?
     end
+  end
+
+  def raise_no_cancelled_invoices
+    return unless @unpaid_invoices_ids.count == @existing_bans_ids.count
+
+    raise Errors::NoCancelledInvoices.new(user.id, domain_name)
+  end
+
+  def generate_ids_for_invoices_and_bans
+    @unpaid_invoices_ids = Invoice.where(user_id: user, status: Invoice.statuses[:cancelled])
+                                  .pluck(:id)
+    @existing_bans_ids = Ban.where(invoice_id: @unpaid_invoices_ids).pluck(:id)
   end
 end
